requirements.txt
pandas>=1.5.0
numpy>=1.23.0,<2.0.0
scikit-learn>=1.1.0
scikit-surprise>=1.1.3
streamlit>=1.20.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
scipy>=1.9.0
pytest>=7.0.0
